some work done to move date to another file
