+++ src/optional/date_class.h
@@ -1,0 +1,16 @@
+//
+//  date_class.h
+//  gravity
+//
+//  Created by Marco Bambini on 10/01/15.
+//  Copyright (c) 2015 CreoLabs. All rights reserved.
+//
+
+#ifndef __DATE_CLASS__
+#define __DATE_CLASS__
+
+#include "gravity_vm.h"
+#include "gravity_core.h"
+
+
+#endif
